dwm: add a few more actions on WinTitle clicks
Such actions include focusing on scroll and warping cursor to
the selected window on left click. The thing left is menuconfig
(open dmenu with a few options on right click and then find the
resulting option in the array and get the function that we need
to perform and it's arguments, for example killclient, moveresize
fakefullscreen, togglefloating, move, and so on)

Also a bug in warp was fixed, when the position of the status
bar was calculated incorrectly and sometimes cursor was not
registered as being above status bar and the cursor was warped
anyways.